updated the stock correction logic

+++ src/main/java/com/bookshop/dto/ItemDto.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 @Builder
 public class ItemDto {
-    private int id;
+    private Long id;
 
-    private int bookId;
+    private Long bookId;
     private int quantity;
 }
